crawler/requirements.txt
requests==2.31.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
selenium==4.20.0
pandas==2.2.2
